Allow node versions >=0.6
Nest actually supports node 0.7 as well, as of version nest v0.1.3.  However,
nest v0.1.3 introduced the engines requirement, which was not present prior.
As a result, node 0.7 users get nest 0.1.2, which *isn't* compatible (due to
use of the "sys" module.)

Recommend bumping the version and re-publishing if this patch is acceptable.
The http apis that nest uses are stable, and backwards compatibility can be
assumed moving forward.  However, if you wish to be a bit more conservative,
then `>=0.6 <0.9` would give you a bit of protection.